+++ Common/changelog.md
@@ -21,6 +21,7 @@
 - Bat AI changes
   - Player bats now flee monsters and have a revenge goal
   - Herobrine bat attacks players
+  - Herobrine bat teleports away when under certain health
   - Make witch bat either have regeneration or resistance when below certain amounnt of health
 - Bat model changes
   - Make pumpkin and undead bats slightly bigger